Backend/requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.7
Django==3.2.9
django-cors-headers==3.10.0
django-environ==0.8.1
django-mysql==4.2.0
djangorestframework==3.12.4
gunicorn==20.1.0
idna==3.3
mysqlclient
Pillow==8.4.0
pytz==2021.3
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
stripe==2.61.0
urllib3==1.26.7
